+++ neighborhood-library/navigation/MainTabNavigator.js
@@ -4,7 +4,7 @@ import { createStackNavigator, createBottomTabNavigator } from 'react-navigation
 
 import TabBarIcon from '../components/TabBarIcon';
 import HomeScreen from '../screens/HomeScreen';
-import LinksScreen from '../screens/LinksScreen';
+import ShelfScreen from '../screens/ShelfScreen';
 import SettingsScreen from '../screens/SettingsScreen';
 
 const HomeStack = createStackNavigator({
@@ -25,11 +25,11 @@ HomeStack.navigationOptions = {
   ),
 };
 
-const LinksStack = createStackNavigator({
-  Links: LinksScreen,
+const ShelfStack = createStackNavigator({
+  Shelf: ShelfScreen,
 });
 
-LinksStack.navigationOptions = {
+ShelfStack.navigationOptions = {
   tabBarLabel: 'Shelf',
   tabBarIcon: ({ focused }) => (
     <TabBarIcon
@@ -83,7 +83,7 @@ MoreStack.navigationOptions = {
 
 export default createBottomTabNavigator({
   HomeStack,
-  LinksStack,
+  ShelfStack,
   SettingsStack,
   BrowseStack,
   MoreStack
