+++ neighborhood-library/package.json
@@ -15,6 +15,7 @@
     "expo": "^32.0.0",
     "react": "16.5.0",
     "react-native": "https://github.com/expo/react-native/archive/sdk-32.0.0.tar.gz",
+    "react-native-paper": "^2.16.0",
     "react-navigation": "^3.0.9"
   },
   "devDependencies": {
